fix compiler warnings for implicit type casts

+++ test/unit/cpp/data_cache.cpp
@@ -73,8 +73,8 @@ BOOST_FIXTURE_TEST_CASE(AddAddr, DataCacheFixture)
 
     BOOST_TEST(AddrInx::item_type == CIT::Address);
     BOOST_TEST(AddrInx::data_type == CDT::cdStr);
-    BOOST_TEST(addr_inx.magic_index == 0x01040001);
-    BOOST_TEST(addr_inx() == 1);
+    BOOST_TEST(addr_inx.magic_index == uint32_t(0x01040001));
+    BOOST_TEST(addr_inx() == uint32_t(1));
     assert_cache_state();
 }
 
@@ -86,8 +86,8 @@ BOOST_FIXTURE_TEST_CASE(AddInt, DataCacheFixture)
     IntInx int_inx = dc.extern_int(&style_coloring);
     BOOST_TEST(IntInx::item_type == CIT::Value);
     BOOST_TEST(IntInx::data_type == CDT::cdInt);
-    BOOST_TEST(int_inx.magic_index == 0x02010000);
-    BOOST_TEST(int_inx() == 0);
+    BOOST_TEST(int_inx.magic_index == uint32_t(0x02010000));
+    BOOST_TEST(int_inx() == uint32_t(0));
     assert_cache_state();
 }
 
@@ -99,8 +99,8 @@ BOOST_FIXTURE_TEST_CASE(AddBool, DataCacheFixture)
     BoolInx bool_inx = dc.extern_bool(&show_footer_db);   // not backed
     BOOST_TEST(IntInx::item_type == CIT::Value);
     BOOST_TEST(IntInx::data_type == CDT::cdInt);
-    BOOST_TEST(bool_inx.magic_index == 0x02030000);
-    BOOST_TEST(bool_inx() == 0);
+    BOOST_TEST(bool_inx.magic_index == uint32_t(0x02030000));
+    BOOST_TEST(bool_inx() == uint32_t(0));
     assert_cache_state();
 }
 
@@ -112,8 +112,8 @@ BOOST_FIXTURE_TEST_CASE(AddFloat, DataCacheFixture)
     FloatInx float_inx = dc.extern_float(&font_scale_main);    // not backed
     BOOST_TEST(FloatInx::item_type == CIT::Value);
     BOOST_TEST(FloatInx::data_type == CDT::cdFloat);
-    BOOST_TEST(float_inx.magic_index == 0x02020000);
-    BOOST_TEST(float_inx() == 0);
+    BOOST_TEST(float_inx.magic_index == uint32_t(0x02020000));
+    BOOST_TEST(float_inx() == uint32_t(0));
     assert_cache_state();
 }
 
@@ -127,12 +127,12 @@ BOOST_FIXTURE_TEST_CASE(AddString, DataCacheFixture)
     backed_str_count = 3;
     BOOST_TEST(StrInx::item_type == CIT::Value);
     BOOST_TEST(StrInx::data_type == CDT::cdStr);
-    BOOST_TEST(str_inx.magic_index == 0x02040002);
-    BOOST_TEST(str_inx() == 2);
+    BOOST_TEST(str_inx.magic_index == uint32_t(0x02040002));
+    BOOST_TEST(str_inx() == uint32_t(2));
     BOOST_TEST(AddrInx::item_type == CIT::Address);
     BOOST_TEST(AddrInx::data_type == CDT::cdStr);
-    BOOST_TEST(addr_inx.magic_index == 0x01040001);
-    BOOST_TEST(addr_inx() == 1);
+    BOOST_TEST(addr_inx.magic_index == uint32_t(0x01040001));
+    BOOST_TEST(addr_inx() == uint32_t(1));
     assert_cache_state();
 }
